refactor: use mobx for state and separate files for components
