+++ app/assets/javascripts/preferences/panes/AccountPreferences.tsx
@@ -1,4 +1,4 @@
-import { Sync, Subscription } from '@/preferences/panes/account';
+import { Sync, SubscriptionWrapper } from '@/preferences/panes/account';
 import { PreferencesPane } from '@/preferences/components';
 import { observer } from 'mobx-react-lite';
 import { WebApplication } from '@/ui_models/application';
@@ -7,7 +7,7 @@ export const AccountPreferences = observer(({application}: {application: WebAppl
   return (
     <PreferencesPane>
       <Sync application={application} />
-      <Subscription application={application} />
+      <SubscriptionWrapper application={application} />
     </PreferencesPane>
   );
 });
